Move analysis tests into dreams/analysis folder
