+++ server/tests/dreams/analysis/deleteAnalysis.test.ts
@@ -1,15 +1,15 @@
 import request from 'supertest';
 import { beforeEach, expect, test } from 'vitest';
-import { wipeDB } from '../setup/wipeDB.js'
-import { server } from '../setup/testServer.js'
-import { userOneCreds, userThreeCreds } from '../users/data.js';
-import {createUser, getAuthHeader} from '../users/utils/userCreation.js'
-import { baseUrl } from './utils.js';
-import { assertSingleError } from '../users/utils/assertErrors.js';
+import { wipeDB } from '../../setup/wipeDB.js'
+import { server } from '../../setup/testServer.js'
+import { userOneCreds, userThreeCreds } from '../../users/data.js';
+import {createUser, getAuthHeader} from '../../users/utils/userCreation.js'
+import { baseUrl } from '../utils.js';
+import { assertSingleError } from '../../users/utils/assertErrors.js';
 import { Types } from 'mongoose';
-import { Dream } from '../../models/dream.model.js';
-import { newDreamData } from './data.js';
-import { options } from '../../interfaces/dream.interfaces.js'
+import { Dream } from '../../../models/dream.model.js';
+import { newDreamData } from '../data.js';
+import { options } from '../../../interfaces/dream.interfaces.js'
 
 // Wipe db and save data
 let userOneAuth: [string, string]
